requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

newspaper3k == 0.2.8
nltk == 3.7
textblob == 0.17.1
